+++ gammapy/cube/tests/test_new.py
@@ -100,7 +100,7 @@ def test_MapMaker(mode,expected):
     pos_SagA = SkyCoord(266.41681663, -29.00782497, unit="deg", frame="icrs")
     energy_axis = MapAxis.from_edges([0.1,0.5,1.5,3.0,10.],name='energy',unit='TeV',interp='log')
     geom = WcsGeom.create(binsz=0.1*u.deg, skydir=pos_SagA, width=15.0, axes=[energy_axis])
-    mmaker = MapMaker(geom, geom, offset_max=6.0 * u.deg, cutout_mode=mode)
+    mmaker = MapMaker(geom, 6.0 * u.deg, cutout_mode=mode)
     obs = [110380, 111140]
 
     for obsid in obs:
@@ -111,7 +111,7 @@ def test_MapMaker(mode,expected):
 
     etrue_axis = MapAxis.from_bounds(0.5, 50.0, 10, name='energy', unit='TeV', interp='log')
     geom_etrue = WcsGeom.create(binsz=0.1*u.deg, skydir=pos_SagA, width=15.0, axes=[etrue_axis])
-    maker = MapMaker(geom, geom_etrue, offset_max=6.0 * u.deg, cutout_mode=mode)
+    maker = MapMaker(geom, offset_max=6.0 * u.deg, ref_geom_etrue=geom_etrue, cutout_mode=mode)
     obslist = ds.obs_list(obs)
     maps = maker.run(obslist)
     assert maps['exposure_map'].unit == "m2 s"
